refactor login and signup pages to use single component
