+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "chessticulate-fe",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
