+++ src/app/login/page.tsx
@@ -1,9 +1,9 @@
-import LoginForm from "@/components/LoginForm";
+import UserForm from "@/components/UserForm";
 
 export default function Login() {
   return (
     <main className="flex justify-center pt-28">
-      <LoginForm />
+      <UserForm />
     </main>
   );
 }
