+++ src/app/signup/page.tsx
@@ -1,9 +1,9 @@
-import SignupForm from "@/components/SignupForm";
+import UserForm from "@/components/UserForm";
 
 export default function Signup() {
   return (
     <main className="flex justify-center pt-28">
-      <SignupForm />
+      <UserForm />
     </main>
   );
 }
